Fix some bug about the merge between Xala's Branch and my branch
