--- src/app/inicio/page.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-export default function Page() {
-    return <h1>Esto es Inicio</h1>
-  }
